+++ docs/changelog.md
@@ -34,11 +34,16 @@ See [upgrade notes][1] for helpful information when upgrading from previous vers
   - App Configuration:
     - Check that replica locations are in allowed regions by @BernieWhite.
       [#3441](https://github.com/Azure/PSRule.Rules.Azure/issues/3441)
+  - Managed Instance for Apache Cassandra:
+    - Check that Managed Instance for Apache Cassandra clusters have availability zones enabled by @BenjaminEngeset.
+      [#3592](https://github.com/Azure/PSRule.Rules.Azure/issues/3592)
   - Cosmos DB:
     - Check that Cosmos DB accounts have availability zones enabled by @BenjaminEngeset.
       [#3055](https://github.com/Azure/PSRule.Rules.Azure/issues/3055)
     - Check that MongoDB vCore clusters use Microsoft Entra ID authentication by @BenjaminEngeset.
       [#3369](https://github.com/Azure/PSRule.Rules.Azure/issues/3369)
+    - Check that MongoDB vCore clusters have availability zones enabled by @BenjaminEngeset.
+      [#3586](https://github.com/Azure/PSRule.Rules.Azure/issues/3586)
   - Data Explorer:
     - Check that public network access is disabled by @BenjaminEngeset.
       [#3114](https://github.com/Azure/PSRule.Rules.Azure/issues/3114)
@@ -52,6 +57,9 @@ See [upgrade notes][1] for helpful information when upgrading from previous vers
   - Application Gateway Policy:
     - Updated `Azure.AppGwWAF.RuleGroups` to use Microsoft Default Rule Set instead of legacy OWASP rule set by @BenjaminEngeset.
       [#3553](https://github.com/Azure/PSRule.Rules.Azure/issues/3553)
+  - Cosmos DB:
+    - Renamed `Azure.Cosmos.DisableLocalAuth` to `Azure.Cosmos.NoSQLLocalAuth` to reflect it applies only to NoSQL API by @BernieWhite.
+      [#3576](https://github.com/Azure/PSRule.Rules.Azure/issues/3576)
 
 ## v1.46.0